Workaround plist security vulnerability in terminal-image
`terminal-image` indirectly depends on `plist 2.0.1` which is vulnerable
to a DOS attack.
We force the use of a more recent version to fix the issue. There is no
incompatibility risk as the breaking change for this library is the drop
of node 4 support.

We also opened directly a bug so that the dependency is updated in the
library that uses this version: https://github.com/sindresorhus/term-img/issues/6