Refactor code in a manner closer to hexagonal architecture spirit
This a very big refator commit (yes bad bad bad !!) whose aim was just to see what
hexagonal-like architecture could be done on this simple project.

This brings a bit more abstraction and inevitably (?) a bit more complexity.
There is now indeed a better layer separation but, unless for the fun here,
it was not probably worth it for such a small script.